requirements.txt
geopy>=1.11.0
protobuf>=3.0.0
flask>=0.11.1
sqlalchemy>=1.1.0
alembic>=0.9.5
#aiopogo>=2.0.2,<2.1
git+https://github.com/Falaris01/aiopogo.git@master#egg=aiopogo
polyline>=1.3.1
aiohttp>=2.1,<2.3
pogeo==0.3.*
cyrandom>=0.1.2
python-crontab>=2.2.4
python-anticaptcha>=0.1.0